Separate wikihero completely
